requirements.txt
# ── Telegram ──────────────────────────────────────────────────────────────────
python-telegram-bot[job-queue]==21.3

# ── Async HTTP ────────────────────────────────────────────────────────────────
aiohttp==3.9.5

# ── Video downloading ─────────────────────────────────────────────────────────
yt-dlp==2024.5.27

# ── MEGA ──────────────────────────────────────────────────────────────────────
# megapy is the actively maintained fork of mega.py
megapy==1.0.9

# ── Utilities ─────────────────────────────────────────────────────────────────
requests==2.32.3          # used internally by some yt-dlp extractors
tenacity==8.3.0           # retry logic (available for future use)
